+++ loop/utils/__init__.py
@@ -1,6 +1,7 @@
 from loop.utils.callbacks import create_callbacks, create_tqdm_callback
 from loop.utils.reporting import format_report_header, format_report_section, format_report_footer
 from loop.utils.generators import generate_permutation
+from loop.utils.splits import split_sequential, split_random
 
 __all__ = [
     'create_callbacks', 
@@ -8,5 +9,7 @@ __all__ = [
     'format_report_header',
     'format_report_section',
     'format_report_footer',
-    'generate_permutation'
+    'generate_permutation',
+    'split_sequential',
+    'split_random'
 ]